% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQB.R
\docType{package}
\name{SQB}
\alias{SQB}
\alias{SQB-package}
\title{Sequential bagging method on Regression}
\description{
Combine bagging, bootstrap decision tree and kernel density estimate
}
\author{
Moshu Xie \email{mxie622@gmail.com}
}
